+++ changelog.md
@@ -8,6 +8,7 @@ Added `LX.toSnakeCase(str)` util.
 CodeEditor:
 - Improved custom suggestions format (now allows for `insertText` too).
 - Supported Multiline TAB indent/dedent.
+- `setText` now has language detection option.
 - Fixed click autocomplete suggestion.
 
 ## 8.3.0 (master)